requirements.txt
# FastAPI & Server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# LangChain (خط 0.3.x)
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
langchain-google-genai==2.0.5
langchain-core==0.3.21

# Vector Database
qdrant-client==1.12.1

# PDF Processing
pypdf==5.1.0

# File Upload Support
python-multipart==0.0.18

# Env Vars
python-dotenv==1.0.1

# Google AI SDK
google-generativeai==0.8.3

# Embeddings & Transformers
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.46.3

# Data Validation
pydantic==2.10.3
# pydantic-core يُثبَّت تلقائيًا مع pydantic

# HTTP Client
httpx==0.28.1

# Core deps aligned with LangChain 0.3.x
numpy==1.26.4
typing-extensions==4.12.2
